fix(mxfp8): int64 expert_id in Triton kernel to avoid >2GB stride overflow
The M3 GPU-prefill fallback (kt_ep_wrapper.py L2562) crashed with
cudaErrorIllegalAddress whenever a hot expert in a 4096+ token prefill
fired the full 128-expert masked GEMM. The crash was not a bound-check
issue (masked_m always fit in m_buf with OVER_M=0) but a plain integer
overflow in the first stride multiplication inside the Triton kernel.

Diagnosis via per-call probe in deep_gemm.py + standalone repro
/tmp/repro_mxfp8_oob.py reproduced reliably at the exact shapes the
dispatch hands to the kernel during a 4096-prefill:
  E=128, M=4096, K=6144 up-gemm   stride_a_e = 25,165,824
  E=128, M=4096, K=384  down-gemm stride_a_e =  1,572,864

tl.program_id(2) returns int32 by default. With expert_id up to 127 and
stride_a_e = 25M, the product 127 * 25M = 3.2G exceeds INT32_MAX = 2.1G.
The overflow wraps to a negative offset; a_base then points well below
the tensor base and the SM90 MMU traps on the first load.

The hybrid 8/128-expert micro-batch path (probes #0..#34 in the log)
never tripped this because 8 * 1.5M = 12M is firmly inside int32; only
the threshold-triggered full-fallback call (probes #35) crashed.

Fix: cast tl.program_id(2) to tl.int64 at kernel entry. All downstream
pointer arithmetic promotes to int64 once expert_id is i64.

Repro shape table:
  (8, 256, 6144)      OK before and after
  (128, 4096, 6144)   CRASHED before, OK after
  (128, 4096, 384)    CRASHED before, OK after

+++ python/sglang/srt/layers/quantization/grouped_mxfp8_kernel.py
@@ -50,7 +50,14 @@ def _grouped_mxfp8_gemm_kernel(
     BLOCK_N: tl.constexpr,
     BLOCK_K: tl.constexpr,
 ):
-    expert_id = tl.program_id(axis=2)
+    # Promote expert_id to int64 before any stride multiplication. With
+    # num_experts=128 and stride_a_e = max_m * K (e.g. 4096*6144 = 25M
+    # elements), expert_id * stride_a_e exceeds 2^31 for expert_id >= 86
+    # and overflows signed int32, wrapping to a negative offset → CUDA
+    # illegal memory access. The earlier hybrid 8-expert micro-batches
+    # never tripped this because 8 * stride was within 2^31. Origin:
+    # bug in the M3 full-GPU prefill fallback path (kt_ep_wrapper.py:2562).
+    expert_id = tl.program_id(axis=2).to(tl.int64)
     actual_m = tl.load(masked_m_ptr + expert_id)
 
     pid = tl.program_id(axis=0)
